+++ src/main/java/fr/florent/model/editeur/layer/TileLayer.java
@@ -3,7 +3,7 @@ package fr.florent.model.editeur.layer;
 import fr.florent.model.editeur.tile.Tile;
 
 public class TileLayer extends Layer<Tile> {
-    public TileLayer(int width, int height) {
+    public TileLayer(double width, double height) {
         super(width, height);
     }
 }
